requirements.txt
streamlit
python-dotenv
langchain-google-genai
langchain-core
Pillow
opencv-python
streamlit
python-dotenv
langchain-google-genai
langchain-core
Pillow
opencv-python
langchain-google-genai
python-multipart
uvicorn
fastapi